backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
python-dotenv
python-multipart
twilio
openpyxl
pandas
passlib[bcrypt]
python-jose[cryptography]
python-multipart
